src/monitoring/requirements.txt
pytest>=6.2.5
axiom-py>=0.3.0
python-dotenv>=0.19.0